Fix crash on sending a std::string larger than 2GB
The socket::send(std::string) function calls send_raw which has a signed integer
as the buffer length. With messages larger than 2GB this causes the size_t
passed to ZMQ to be a large positive value near
std::numeric_limits<size_t>::max(). This is fixed by changing send_raw to take
a size_t instead of an int for the buffer length.

+++ src/zmqpp/socket.hpp
@@ -215,7 +215,7 @@ public:
 	 * \param flags message send flags
 	 * \return true if message part sent, false if it would have blocked
 	 */
-	bool send_raw(char const* buffer, int const length, int const flags = normal);
+	bool send_raw(char const* buffer, size_t const length, int const flags = normal);
 
 	/*!
 	 * \warning If the buffer is not large enough for the message part then the
